[TLC-1202] Move getTypeBindRelations to util function
